requirements.txt
anyio
archspec
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
attrs
babel
bash_kernel
beautifulsoup4
bleach
boltons
Brotli
certifi
cffi
chardet
charset-normalizer
click
colorama
comm
contourpy
cycler
debugpy
decorator
defusedxml
distro
dnspython
exceptiongroup
execnet
executing
expecttest
faster-coco-eval
fastjsonschema
filelock
fonttools
fqdn
frozendict
fsspec
gdown
h11
h2
hpack
httpcore
httpx
hyperframe
hypothesis
idna
importlib_resources
iniconfig
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-archive
jupyter-console
jupyter-events
jupyter-http-over-ws
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kagglehub
kiwisolver
libarchive-c
lief
lightning-utilities
lintrunner
mamba
MarkupSafe
matplotlib
matplotlib-inline
mistune
more-itertools
mpmath
nbclient
nbconvert
nbformat
nbzip
nest-asyncio
networkx
ninja
notebook
notebook_shim
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-contrib-python-headless
optree
overrides
packaging
pandas
pandocfilters
parso
pexpect
pickleshare
pillow
pkginfo
pkgutil_resolve_name
platformdirs
plotly
pluggy
prometheus_client
prompt_toolkit
psutil
ptyprocess
pure_eval
py-cpuinfo
pycocotools
pycosat
pycparser
Pygments
pyparsing
PySocks
pytest
pytest-shutil
python-dateutil
python-etcd
python-json-logger
pytz
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
ruamel.yaml
ruamel.yaml.clib
scikit-learn
scipy
seaborn
Send2Trash
six
sniffio
sortedcontainers
soupsieve
stack-data
sympy
tenacity
termcolor
terminado
threadpoolctl
tinycss2
torch
torchaudio
torchelastic
torchmetrics
torchvision
tornado
tqdm
traitlets
triton
truststore
types-dataclasses
types-python-dateutil
typing_extensions
tzdata
ultralytics
ultralytics-thop
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
widgetsnbextension
xml-python
zipp
zstandard
